backend/src/sandbox/agent_infra_sandbox/langchain_tools/pyproject.toml
[project]
name = "langchain-sandbox-tools"
version = "0.1.0"
description = "Production-ready LangChain tools for AIO Sandbox integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Agent Infra Team"}
]
keywords = ["langchain", "sandbox", "ai-agent", "tools", "browser", "terminal", "jupyter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "agent-sandbox>=1.0.0",
    "langchain-core>=0.2.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["langchain_tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
